% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6Methods.R
\name{Bottleneck_size_Calculation}
\alias{Bottleneck_size_Calculation}
\title{Calculate transmission bottleneck size}
\usage{
Bottleneck_size_Calculation(
  transmission_ob,
  method = "KL",
  plot = FALSE,
  show_table = FALSE,
  transmission_pairs = NULL,
  donor_depth_threshold = 500,
  recipient_depth_threshold = 500,
  error_calling = 0.01,
  log = FALSE,
  variant_calling = 0.03,
  Nbmin = 1,
  Nbmax = 1000,
  NonSyn_or_Syn = "All"
)
}
\arguments{
\item{transmission_ob}{is a transmission object created by,`CreateTransmissionObject`.}

\item{method}{is characters containing `KL` method, `Presence-Absence` method, `Binomial` method, `Beta_binomial_Approximate` method, `Beta_binomial_Exact` method and `Wright-Fisher` method}

\item{plot}{is a logical value to determine plot the likelihood or not. Each plot would be stored in an individual folder named by transmission id.}

\item{show_table}{is a logical value to determine output a result table as `csv` format.}

\item{transmission_pairs}{is a dataframe which is the subset of the transmission pairs table during the object creation.}

\item{donor_depth_threshold}{is a integer to filter variants in donor with the lower sequencing coverage}

\item{recipient_depth_threshold}{is a integer to filter variants in recipient with the lower sequencing coverage}

\item{error_calling}{is a numeric and filter. The proportion of the variant sites lower than error calling means the variant is difficult to distinguish from PCR errors. Then, the sites with lower proportion than error calling would be filtered.}

\item{log}{is a logical value to show the log or not. The log would be stored in an individual folder.}

\item{variant_calling}{is a parameter used for `Presence-absence`method, `Binomial` method, `Beta_binomial` method to determine the variants is absent in recipient or not. Proportion of variant site lower than variant calling means this site is absent in recipient.}

\item{Nbmin}{is a integer representing the minimum number in likelihood range.}

\item{Nbmax}{is a integer representing the maximum number in likelihood range.}

\item{NonSyn_or_Syn}{is a filter to keep the variant sites which meet the requirements. If user need the mutation of variant sites are synonymous mutation, then `NonSyn_or_Syn="Synonymous"`. If user need the mutation of variant sites are non-synonymous mutation, then `NonSyn_or_Syn="Non-Synonymous"`. The default is "All"}
}
\value{
Bottleneck size table containing the transmission pairs, bottleneck size, the lower confidence interval and higher confidence interval
}
\description{
Calculate transmission bottleneck size
}
\examples{
#The example can not be run directly
#Use different methods for estimation
#use KL method
KL_table = Bottleneck_size_calculation(transmission_object, method="KL")
#use presence-absence method
PA_table = Bottleneck_size_calculation(transmission_object, method="Presence-Absence")
#use Binomial method
Bi_table = Bottleneck_size_calculation(transmission_object, method="Binomial")
#use Beta-binomial approximate method
BB_App_table = Bottleneck_size_calculation(transmission_object, method="Beta_binomial_Approximate")
#use Beta-binomial approximate method
BB_Exact_table = Bottleneck_size_calculation(transmission_object, method="Beta_binomial_Exact")
#use wright-fisher method
BB_WF_table = Bottleneck_size_calculation(transmission_object, method="Wright-Fisher")
Take KL method as example to show other arguments
#Execute depth thresholds argument. The default of depth thresholds are 500.
KL_table = Bottleneck_size_calculation(transmission_object, method="KL"
                                  ,donor_depth_threshold=1000, recipient_depth_threshold=1000)
#Execute error calling argument. The default of error calling threshold is 0.01.
KL_table = Bottleneck_size_calculation(transmission_object, method="KL",error_calling=0.02)
#Execute range argument. The default of Nbmin is 1 and Nbmax=1000. They are used for
#                                                    methods excluding `Wright-Fisher` method.
KL_table = Bottleneck_size_calculation(transmission_object, method="KL",Nbmin=50,Nbmax=100)
#Execute variant calling argument. The argument only use
#                      for `Presence-Absence`method, `Binomial` method, `Beta_binomial` method.
BB_Exact_table = Bottleneck_size_calculation(transmission_object,
                                               method="Beta_binomial_Exact",variant_calling=0.02)
#if user have the subset of the transmission pairs, `subset_transmissionpairs`which is the
#                                                same dataframe as the input for `CreateTransmissionObject`.
KL_table = Bottleneck_size_calculation(transmission_object, method="KL"
                                                         ,transmission_pairs=subset_transmissionpairs)
#If user need to output likelihood curve of transmission pairs in working directory,
#                                                            excluding `Wright-Fisher` method.
KL_table = Bottleneck_size_calculation(transmission_object, method="KL",plot=TRUE)
#If user need to output table of transmission pairs in working directory.
KL_table = Bottleneck_size_calculation(transmission_object, method="KL",show_table=TRUE)
#If user need to output log of transmission pairs in working directory.
KL_table = Bottleneck_size_calculation(transmission_object, method="KL",log=TRUE)
}
