% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3ObjectCreation.R
\name{CreateTransmissionObject}
\alias{CreateTransmissionObject}
\title{Create Transmission Object}
\usage{
CreateTransmissionObject(transmission_pairs_table)
}
\arguments{
\item{transmission_pairs_table}{is a dataframe only contatining 2 columns, Donor and Recipient}
}
\value{
Transmission_object, a environment containing all the infomation of transmission pairs
}
\description{
This function could check the input files and catch the files in working directory according to transmission
 pairs table to create a transmission object.
}
\examples{
mytransmission_object = CreateTransmissionObject(transmission_pairs_table)

}
