% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6Methods.R
\name{Summary_ob}
\alias{Summary_ob}
\title{Summary_ob}
\usage{
Summary_ob(transmission_ob, save = FALSE, file_name = NULL)
}
\arguments{
\item{transmission_ob}{is a object created by `CreateTransmissionObject` function}

\item{save}{is a logical value to determine whether the table need to be saved in path or not.}

\item{file_name}{is the name of file. It only would be used when `save=TRUE`.}
}
\value{
would return a csv file containing the transmission pairs: donors, recipients and shared variant sites.
}
\description{
Summary_ob
}
\examples{
#The example can not be run directly
summary_table = Summary_ob(transmission_object)
}
